# swagger-output.json
{"openapi":"3.0.0","info":{"title":"CallGenZ API Workspace","version":"1.0.0","description":"Interactive production-ready API documentation for the CallGenZ Backend. Includes authentication, user management, todos, and calendar sync.","contact":{"name":"Backend Architect Team","email":"support@callgenz.app"}},"servers":[{"url":"/api","description":"Dynamic Host (Current URL)"},{"url":"http://localhost:3000/api","description":"Local Development Server"}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT","description":"Standard JWT Authorization header using the Bearer scheme. Enter your token below."}},"schemas":{"User":{"type":"object","properties":{"id":{"type":"string","format":"uuid","example":"f47ac10b-58cc-4372-a567-0e02b2c3d479"},"phone":{"type":"string","example":"+1234567890"},"name":{"type":"string","example":"John Doe"},"age":{"type":"integer","example":30},"createdAt":{"type":"string","format":"date-time"}}},"Todo":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"userId":{"type":"string","format":"uuid"},"title":{"type":"string","example":"Buy Groceries"},"description":{"type":"string","example":"[{\"text\":\"Milk\",\"done\":false}]"},"dueAt":{"type":"string","format":"date-time"},"completed":{"type":"boolean","example":false},"createdAt":{"type":"string","format":"date-time"}}},"Reminder":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"todoId":{"type":"string","format":"uuid"},"callTime":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["PENDING","CALLED","MISSED","COMPLETED"],"example":"PENDING"},"retryCount":{"type":"integer","example":0},"createdAt":{"type":"string","format":"date-time"}}},"ErrorResponse":{"type":"object","properties":{"status":{"type":"string","example":"fail"},"message":{"type":"string","example":"Detailed error message here"}}}},"responses":{"UnauthorizedError":{"description":"Access token is missing, invalid, or expired","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"NotFoundError":{"description":"The requested resource was not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"BadRequestError":{"description":"Invalid input syntax or missing required fields","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"security":[{"bearerAuth":[]}],"paths":{"/auth/send-otp":{"post":{"summary":"Send OTP to phone number (Legacy Twilio)","tags":["Auth"],"security":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["phone"],"properties":{"phone":{"type":"string","example":"+919876543210"}}}}}},"responses":{"200":{"description":"OTP sent successfully"},"400":{"$ref":"#/components/responses/BadRequestError"}}}},"/auth/verify-otp":{"post":{"summary":"Verify Twilio OTP and login","tags":["Auth"],"security":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["phone","otp"],"properties":{"phone":{"type":"string","example":"+919876543210"},"otp":{"type":"string","example":"123456"}}}}}},"responses":{"200":{"description":"Auth success","content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5c..."},"user":{"$ref":"#/components/schemas/User"}}}}}},"400":{"$ref":"#/components/responses/BadRequestError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/auth/firebase-login":{"post":{"summary":"Verify Firebase IdToken and securely login","tags":["Auth"],"security":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["firebaseToken"],"properties":{"firebaseToken":{"type":"string","description":"IdToken retrieved from Firebase Auth frontend SDK"}}}}}},"responses":{"200":{"description":"Auth success","content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string","description":"CallGenZ backend JWT for subsequent requests","example":"eyJhbGciOiJIUzI1NiIsInR5c..."},"user":{"$ref":"#/components/schemas/User"}}}}}},"400":{"$ref":"#/components/responses/BadRequestError"},"401":{"$ref":"#/components/responses/UnauthorizedError"},"500":{"description":"Firebase Admin not initialized properly on server"}}}},"/calendar":{"post":{"summary":"Create a deadline-driven calendar task","tags":["Calendar"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["title","dueAt"],"properties":{"title":{"type":"string","example":"Dentist Appointment"},"description":{"type":"string"},"dueAt":{"type":"string","format":"date-time","example":"2026-10-15T10:00:00.000Z"}}}}}},"responses":{"201":{"description":"Task created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}},"400":{"$ref":"#/components/responses/BadRequestError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}},"get":{"summary":"Retrieve all assigned calendar tasks","tags":["Calendar"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Array of mapped calendar tasks","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Todo"}}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/calendar/{id}":{"get":{"summary":"Retrieve a single calendar task","tags":["Calendar"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Task specific metadata","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"404":{"$ref":"#/components/responses/NotFoundError"}}},"put":{"summary":"Modify an existing calendar task","tags":["Calendar"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"completed":{"type":"boolean"},"dueAt":{"type":"string","format":"date-time"}}}}}},"responses":{"200":{"description":"Task updated state","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}},"400":{"$ref":"#/components/responses/BadRequestError"},"401":{"$ref":"#/components/responses/UnauthorizedError"},"404":{"$ref":"#/components/responses/NotFoundError"}}},"delete":{"summary":"Remove a scheduled calendar task","tags":["Calendar"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Task successfully deleted","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"Task deleted"}}}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"404":{"$ref":"#/components/responses/NotFoundError"}}}},"/notifications/tokens":{"post":{"summary":"Register a new device push token","tags":["Notifications"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["token"],"properties":{"token":{"type":"string","description":"The unique FCM token generated from the client device","example":"eXunG_dFk...R8A2B8"}}}}}},"responses":{"201":{"description":"Push token registered successfully","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"Token saved"}}}}}},"400":{"$ref":"#/components/responses/BadRequestError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}},"get":{"summary":"Retrieve all active push tokens for the user","tags":["Notifications"],"responses":{"200":{"description":"Array of registered device tokens","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"token":{"type":"string"},"userId":{"type":"string","format":"uuid"}}}}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/notifications/tokens/{id}":{"delete":{"summary":"Unregister a specific push token","tags":["Notifications"],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Token deleted successfully","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"Token removed"}}}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"404":{"$ref":"#/components/responses/NotFoundError"}}}},"/reminders":{"post":{"summary":"Schedule a new call reminder linked to a Todo","tags":["Reminders"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["todoId","callTime"],"properties":{"todoId":{"type":"string","format":"uuid","description":"ID of the Todo item to remind the user about"},"callTime":{"type":"string","format":"date-time","example":"2026-10-15T10:00:00.000Z"}}}}}},"responses":{"201":{"description":"Reminder scheduled successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reminder"}}}},"400":{"$ref":"#/components/responses/BadRequestError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}},"get":{"summary":"Retrieve history and upcoming call reminders","tags":["Reminders"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Array of reminder objects","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Reminder"}}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/reminders/{id}":{"get":{"summary":"Get metadata of a specific reminder","tags":["Reminders"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Reminder details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reminder"}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"404":{"$ref":"#/components/responses/NotFoundError"}}},"put":{"summary":"Update lifecycle state or time of a scheduled reminder","tags":["Reminders"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["PENDING","CALLED","MISSED","COMPLETED"],"description":"Current status of the reminder"},"callTime":{"type":"string","format":"date-time","description":"New scheduled call time"},"retryCount":{"type":"integer","description":"Number of retries for the call"}}}}}},"responses":{"200":{"description":"Reminder updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reminder"}}}},"400":{"$ref":"#/components/responses/BadRequestError"},"401":{"$ref":"#/components/responses/UnauthorizedError"},"404":{"$ref":"#/components/responses/NotFoundError"}}},"delete":{"summary":"Cancel a scheduled reminder","tags":["Reminders"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Reminder cancelled and deleted","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"Reminder deleted"}}}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"404":{"$ref":"#/components/responses/NotFoundError"}}}},"/todos":{"post":{"summary":"Create a new todo item","tags":["Todos"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["title"],"properties":{"title":{"type":"string","example":"Buy Groceries"},"description":{"type":"string","description":"Stringified JSON array of subtasks","example":"[{\"text\":\"Milk\",\"done\":false}]"},"dueAt":{"type":"string","format":"date-time"}}}}}},"responses":{"201":{"description":"Todo created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}},"400":{"$ref":"#/components/responses/BadRequestError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}},"get":{"summary":"Retrieve all user's todos","tags":["Todos"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"List of todos","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Todo"}}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/todos/{id}":{"get":{"summary":"Retrieve a specific todo by ID","tags":["Todos"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Todo details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"404":{"$ref":"#/components/responses/NotFoundError"}}},"put":{"summary":"Update a specific todo","tags":["Todos"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string","description":"Stringified JSON array of subtasks"},"completed":{"type":"boolean"},"dueAt":{"type":"string","format":"date-time"}}}}}},"responses":{"200":{"description":"Updated todo item","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"404":{"$ref":"#/components/responses/NotFoundError"}}},"delete":{"summary":"Delete a specific todo","tags":["Todos"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Todo deleted successfully","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"Todo deleted"}}}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"404":{"$ref":"#/components/responses/NotFoundError"}}}},"/users/me":{"get":{"summary":"Retrieve current logged-in user profile","tags":["Users"],"responses":{"200":{"description":"User profile data retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"404":{"$ref":"#/components/responses/NotFoundError"}}},"put":{"summary":"Update current user profile details","tags":["Users"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"Jane Doe"},"age":{"type":"integer","example":28}}}}}},"responses":{"200":{"description":"User profile updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"$ref":"#/components/responses/BadRequestError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}},"delete":{"summary":"Delete current user account and all associated data","tags":["Users"],"responses":{"200":{"description":"Account deleted successfully","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"User deleted"}}}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}}},"tags":[{"name":"Auth","description":"Phone OTP and Firebase authentication"},{"name":"Calendar","description":"Calendar events and deadline scheduling"},{"name":"Notifications","description":"FCM Push Token registration and device tracking"},{"name":"Reminders","description":"Automated phone call reminders scheduling"},{"name":"Todos","description":"Task management and checklists"},{"name":"Users","description":"User profile and identity management"}]}